requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
cachetools==4.2.2
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.6
django-common-helpers==0.9.2
django-cron==0.5.1
django-crontab==0.7.1
django-environ==0.4.5
django-extensions==3.1.3
django-redis==5.0.0
djangorestframework==3.12.4
google-api-core==1.31.2
google-api-python-client==2.15.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
kombu==5.1.0
netifaces==0.11.0
packaging==21.0
prompt-toolkit==3.0.19
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
redis==3.5.3
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
